Only update database if announcements in YAML file
We want to make sure that we don't accidentally overwrite the existing
announcements if the YAML doesn't contain any entries i.e. if the task
is run again after the announcements have been migrated.

+++ app/services/coronavirus_pages/announcements_builder.rb
@@ -1,14 +1,16 @@
 class CoronavirusPages::AnnouncementsBuilder
   def create_announcements
-    Announcement.transaction do
-      coronavirus_page.announcements.delete_all
+    if announcements_from_yaml.any?
+      Announcement.transaction do
+        coronavirus_page.announcements.delete_all
 
-      announcements_from_yaml.each do |announcement|
-        coronavirus_page.announcements.create!(
-          text: announcement["text"],
-          href: announcement["href"],
-          published_at: Date.parse(announcement["published_text"]),
-        )
+        announcements_from_yaml.each do |announcement|
+          coronavirus_page.announcements.create!(
+            text: announcement["text"],
+            href: announcement["href"],
+            published_at: Date.parse(announcement["published_text"]),
+          )
+        end
       end
     end
   end
@@ -16,8 +18,8 @@ class CoronavirusPages::AnnouncementsBuilder
 private
 
   def announcements_from_yaml
-    github_data = YamlFetcher.new(coronavirus_page.raw_content_url).body_as_hash
-    github_data["content"]["announcements"]
+    @github_data ||= YamlFetcher.new(coronavirus_page.raw_content_url).body_as_hash
+    @github_data["content"]["announcements"]
   end
 
   def coronavirus_page
